Add 2015 leap second
I noticed that when calculating dates after the last leap second in the
table, a user warning is issued. This is good, as I otherwise would not
have known that leap seconds are handled as a constant.

This patch adds the 2015 leap second to the constant list, bringing it
up to date.